# Conflicts:
#	src/execution/index/art/const_prefix_handle.cpp
#	src/include/duckdb/execution/index/art/art.hpp

+++ scripts/regression/test_runner.py
@@ -50,6 +50,9 @@ parser.add_argument(
     "--clear-benchmark-cache", action="store_true", help="Clear benchmark caches prior to running", default=False
 )
 parser.add_argument(
+    "--keep-benchmark-data", action="store_true", help="Benchmark data will not be deleted between tests", default=False
+)
+parser.add_argument(
     "--regression-threshold-seconds",
     type=float,
     default=0.05,
@@ -71,6 +74,8 @@ disable_timeout = args.disable_timeout
 max_timeout = args.max_timeout
 root_dir = args.root_dir
 no_summary = args.no_summary
+clear_benchmark_cache = args.clear_benchmark_cache
+keep_benchmark_data = args.keep_benchmark_data
 regression_threshold_seconds = args.regression_threshold_seconds
 
 
@@ -158,7 +163,7 @@ if not os.path.isfile(new_runner_path):
     print(f"Failed to find new runner {new_runner_path}")
     exit(1)
 
-if args.clear_benchmark_cache:
+if clear_benchmark_cache:
     old_cache_path = os.path.join(os.path.dirname(old_runner_path), '..', '..', '..', 'duckdb_benchmark_data')
     new_cache_path = os.path.join(os.path.dirname(new_runner_path), '..', '..', '..', 'duckdb_benchmark_data')
     try:
@@ -284,8 +289,9 @@ else:
     print(f"New timing geometric mean: {time_b}")
 
 # nuke cached benchmark data between runs
-if os.path.isdir("duckdb_benchmark_data"):
-    shutil.rmtree('duckdb_benchmark_data')
+if not keep_benchmark_data:
+    if os.path.isdir("duckdb_benchmark_data"):
+        shutil.rmtree('duckdb_benchmark_data')
 
 if summary and not no_summary:
     print(
